Move new JSON config test to BQ system tests file
